Fixed several bugs discovered by running FindBugs (potential null references, comparing string pointers instead of contents, etc.)
Fixed null reference bug introduced in revision 3691 (connection timeout 
was specified before connection was created).



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@3709 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genometryImpl/src/com/affymetrix/genometryImpl/parsers/gchp/AffySingleChromData.java
@@ -79,7 +79,7 @@ public final class AffySingleChromData {
 
 	void skip(DataInputStream dis) throws IOException {
 		int rowSize = totalRowSize();
-		long skipSize = rowCount * rowSize;
+		long skipSize = (long)rowCount * rowSize;	// cast to long before multiplying to avoid possible overflow
 
 		while (skipSize > 0) {
 			long skipped = dis.skip(skipSize);
